move /palettes out of index.js
